+++ HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ---------------
 
-Current
+1.2.2 (2018-06-28)
 ++++++++++++++++++
 
 **Improvements**